Automatically show WebViews created via 'WebContentsCreated'
This is to be consistent with the behavior prior to the switch to Aura.  In
pre-Aura, the WebContentsView HWND was visible by default.  However, in Aura,
the NativeWidget HWND is invisible by default.  This causes WebViews created
using 'window.open' to be invisible because applications never needed to call
'show()' in the past.  So let's just call 'show()' internally so that they
don't have to.