Treat focus changes as activation changes for child windows
When integrating with Aura [1], it was decided to use DesktopNativeWidgetAura
(with the WS_CHILD style added), instead of using NativeWidgetAura, because the
desktop version seemed more complete, and it is what is actually used by
content shell.

However, all the focus stuff for DesktopNativeWidgetAura is handled in the
window activation codepath.  This caused focus to be broken for us because our
child HWND was receiving WM_SETFOCUS and WM_KILLFOCUS, instead of WM_ACTIVATE.

This commit solves this issue by treating WM_SETFOCUS and WM_KILLFOCUS as
window activation changes whenever the DesktopNativeWidgetAura is really a
child window.

[1] 1433f6d "Fix compile errors for v32, integrate with Aura"