Re-expose custom non-client hit-test/dragging in Aura
In previous commit [1,2], we added the ability for blpwtk2-based applications
to implement their own non-client hit-test and dragging.  This was done by
hooking into the WM_NC* handlers in WebContentsViewWin and forwarding those
messages out to the WebViewDelegate.  Those changes were lost in the conversion
to Aura, so this commit restores those hooks from HWNDMessageHandler.

Note that HWNDMessageHandler and its delegate already have a whole mechanism
for handling non-client messages (e.g. GetNonClientComponent), but it requires
a bunch of things like the frame-type to be correctly specified up-front.

In our case, all we want to do is forward these messages to the application and
let it handle them in whatever way it wants, regardless of how the widget was
configured in the beginning.  This commit just tunnels those messages through
the various layers: HWNDMessageHandler -> DesktopRootWindowHostWin -> Widget ->
NativeViewWidget -> NativeViewWidgetDelegate (blpwtk2::WebViewImpl).

Note that these messages are now exposed through NativeViewWidgetDelegate
instead of WebContentsDelegate.

[1] 57efad7 "Support NC dragging in RENDERER_MAIN thread mode"
[2] 63ec22f "Move NC drag support out to WebViewDelegate"