Adjust ffmpegsumo product name to be based on bb_version
This allows multiple versions of ffmpegsumo to coexist, just like all the other
versioned blpwtk2 dlls.  This requires a change to the .lib generator in
generate_stubs.py to make it accept the module name on the command line (from
gyp), instead of deriving the module name from the sigs filename.