Propagate browser command-line switches to renderers in client processes
In a previous commit [1], the command-line for in-process renderers in blpwtk2
client processes was fixed to include the 'kLang' and gpu switches.  This fixed
a bunch of gpu-related issues.

In this commit, we go one step further and propagate the browser command-line
switches as well.  This is similar to how browser switches are propagated to
the renderers launched by RenderProcessHostImpl.

To do this, we reuse a bunch of code in RenderProcessHostImpl, but make it
static so that we don't need a RenderProcessHostImpl instance to propagate
these switches.  The only switch that needs a RenderProcessHostImpl instance
(actually, the browser-context in the RenderProcessHostImpl) is
kDisableDatabases.  This can be handled by blpwtk2 in a future commit.

The 'AppendRendererCommandLine' method that is used by RenderProcessHostImpl
when launching its own renderers now just calls the static methods, then
handles the kDisableDatabases switch for incognito profiles.  It also does
'ContentBrowserClient::AppendExtraCommandLineSwitches', which is actually just
a no-op for us.

[1] cefff91 "Fix command-line for in-process renderers in client processes"