Let application main thread know about WebView's HWND
This is to fix a destruction race condition that happens when the RENDERER_MAIN
thread-mode is used.  In this mode, the application thread is different from
the browser UI thread.  When the application's window is destroyed, Windows
internally destroys all child windows, including the WebView's HWND in the
browser UI thread.  This causes all kinds of problems.

To solve this, we will let the WebViewProxy know about the WebView's HWND, and
also the singleton 'ui::GetHiddenWindow'.  Now 'WebViewProxy::destroy' can
reparent the WebView's HWND into the hidden window in the application's main
thread, before the application's parent window is destroyed.  This prevents
Windows from destroying the WebView's HWND.