Port changes from FontChromiumWin to FontHarfBuzz
Due to the switch to use HarfBuzz [1], the changes from FontChromiumWin.cpp did
not get included in the previous merge.  This commit adds those changes to
FontHarfBuzz.cpp.

Unfortunately, although the font-smoothing flag is now being honored, the LCD
text rendering does not work on canvas elements anymore, even when rendering on
solid backgrounds.  This issue should be revisited in the future.

[1] https://codereview.chromium.org/64473005