Briefly release capture while setting focus
The aura focus controller has some logic that prevents the window that has
capture from also getting focus [1].  It's not clear what the reasoning behind
that is, but it is preventing us from taking focus while handling the
mouse-down event.

This commit works around this issue by temporarily releasing capture if we have
capture due to the mouse being down.

[1] https://github.com/bloomberg/chromium.bb/blob/master/src/ui/views/corewm/focus_controller.cc#L136