Restore static_library versions of libEGL and libGLESv2
The previous merge lost some gyp changes that were made to angle in [1].  This
commit restores those changes.

Also, because the source list is determined using the 'enumerate_files.py'
script, this commit adds a new feature to the script to be able to exclude
certain files (i.e. the files that should be in the shared_library, not in the
static_library).

[1] cd4217d "Add blpangle target"