Send resize ack if renderer's expected size is different from its actual size
In RENDERER_MAIN thread mode, blpwtk2 sends the new size info to the render
widget host as well as directly change the size of the render widget. This
allows the size of the view to change more synchronously.

The problem is that the render widget's size is now different than what the
host expects and this inconsistency causes the host to wait forever for an
ACK from the render widget.

This problem is alleviated by making the render widget keep track of what the
host thinks its size should be. If the host attempts to resize the render
widget to its already set size, the render widget will ACK the resize only
if it knows that there is an inconsistency between the render widget and the
host.