Make V8 use the same ICU data as content
In a previous commit [1], we added code to initialize ICU in V8 separately from
base, because we build V8 as a separate shared library.  Since we are now using
a data file for ICU data instead of a DLL, this means we would have two copies
of the ICU data loaded in memory.

This commit fixes this by extending the V8 ICU initialization API to accept the
raw ICU data, instead of loading the data from a file.  This data pointer is
obtained from the ICU data loaded by base.

[1] 6b7a763 "Fix V8 ICU initialization"