Ensure that ScriptCompiler::CachedData is allocated/deallocated within V8
This is necessary because we run V8 in a separate dll heap, which means all V8
objects need to be allocated/deallocated within V8.

This commit makes the constructors and destructor for CachedData private so
that they cannot be used from outside.  All creation/destruction must go
through the static create() and dispose() functions.

This means we cannot use OwnPtr<v8::ScriptCompiler::CachedData> (unless we
provide a custom deleter).  However, since the OwnPtr in V8ScriptRunner is
leaked right after it is created, we don't actually need to use OwnPtr in this
case.