Add the activateWindowOnMouseDown flag in WebViewCreateParams
Prior to this commit, the setKeyboardFocusOnMouseDown flag was used to
determine if the owner window should be activated on mouse down. This commit
adds a new flag that is only used to determine if the owner window should be
activated on mouse down. This way, the setKeyboardFocusOnMouseDown flag is not
dual-purposed.