Make sure an overridden height is set by the parent flex box
Update the overridden height if it isn't already set. This is needed for
nested row-based flexboxes.  When the content of the inner flexbox causes a
relayout, the inner flexbox will update its height based on its contents. If
the height is set by the parent, the relayout will consider the allocated
space provided by the parent flexbox.