Set the correct flag to indicate the message was handled
When the window does not have non-client area, the 'OnNCHitTest' handler calls
SetMsgHandled(FALSE).  This happens within the mousewheel handler, which
internally performs a hit-test.  This causes the mousewheel event to be treated
as unhandled, causing the mousewheel event to be sent to the parent.

This commit fixes the issue by ensuring that we explicitly call SetMsgHandled
to reset the flag to what it should be.