Ensure managers are initialized at ChildThread construction time
In a previous commit [1], we modified ChildThread to defer creation of the
channel until a later time when the channel name was determined.  This was done
by moving the channel initialization code from the constructor to a separate
InitChannel method.

In the recent merge, there was some code added upstream to initialize some
manager objects in ChildThread.  Since this was added at the end of the
constructor upstream, they ended up at the end of the InitChannel method
downstream.  This was causing crashes when the manager is used and the channel
hasn't been established yet.

This commit splits the manager initialization code from InitChannel into a
separate "InitManagers" method, which is called at the end of the constructor.

[1] a0c5e3e "Always start the render thread"