Check peer of InProcessResourceLoaderBridge after sending headers
In some cases, the InProcessResourceLoaderBridge is disposed as soon as the
headers are sent (this can happen when there is an error response, for
example).  So we need to check that the peer is still valid before attempting
to use it.

Also, fix a potential case where the embedder's reference to the resource
context is not released when the request is canceled and finished at the same
time.