Added setConsoleLogMessageHandler to ToolkitCreateParams
This method allows the embedder to install a callback that will be invoked
when a message is added to the web console. The callback includes
additional information that is specific to the web console, such as the
column number of the script code and optionally the stack trace of the
function that added the message to the console. The stack trace is included
only if the severity of the message is that of an error or higher.