Enable LogMessage crash handling even if wtk2 log message handler is used
In a previous commit [1], we added the ability to capture log messages in wtk2.
If wtk2 handled the message, LogMessage would do no further processing.
However, we still want FATAL messages to crash the application.  This commit
changes LogMessage to continue after the wtk2 handler took care of it, but only
for FATAL messages.  If wtk2 handled the message, then we explicitly don't log
to the system debug log or to the file.

[1] 1af29c1 "Expose an API to get first dibs on log messages"