Remove usage of WebContentsObserver::AboutToNavigateRenderFrame
We previously used this callback to be notified about what the current
RenderView routing-id is, so that we can lookup the RenderView on the
WebViewProxy side.

However, there are cases where this callback is not invoked reliably.  So
replace this callback with RenderViewCreated and RenderViewHostChanged.  Also,
rename rendererRoutingId to renderViewRoutingId to be more precise about the
terminology.

Eventually, RenderView will be removed (see site-isolation discussions
upstream), and we'll have to replace our usage of RenderView with something
else.  But we can cross that bridge when we get to it.