Provide functions that create V8 contexts that can access blink DOM state
This commit extends both blpwtk2 and blink to create V8 contexts that aren't
themselves associated with a particular DOM hierarchy or browser window but
nonetheless need to access and modify contexts that are.

Chromium/blink requires that V8 contexts that touch any DOM state maintain a
reference to a static data structure called the "DOM main world" (the main
world is shared by all browser DOM hierarchies and isn't associated with a
single borwser window).  This reference is stored in an instance of class
ScriptState which is part of a context's embedder data; contexts that lack this
information and attempt to access DOM objects will trigger exceptions with
obscure explanations ("Not a Smi") or will fail assertions in blink.

This commit adds the function blink::WebScriptContext::create() which
constructs a new V8 context with this additional state.  This function is
exposed through blpwtk2::Toolkit::createWebScriptContext().

This commit also makes changes to the functions in v8::BindingSecurity that
perform security checks when accessing data structures.
Settings::webSecurityEnabled() is consulted to allow access to DOM data
structures from V8 contexts that don't maintain a DOM window object.