Don't log channel error after final sync
This is to workaround a problem where OnChannelError() is called in the
ProcessHost when the ProcessClient closes the channel.  This is because a 109
Windows error happens ("The pipe has been ended.").

However, if the ProcessClient is gracefully shutting down, we don't want to log
an error on the host side.  We use the existing Sync message to tell the host
that this is the final sync, so don't bother logging error messages after this.