drawContents() now accept a CSS class list
This class list is temporarily applied to the body element of the document
during the draw operation. This is slightly different than CSS media selectors
because it allows the caller to specify a list of classes, instead of just a
boolean state of whether the draw operation is printing related.

This commit also includes the following changes:
 - in addition to destination width & height, drawContents now requires the
   caller to specify the top-left position of the destination. This can be set
   to 0,0 to get the same affect but it allows the caller to offset the draw
   operation if needed.
 - drawContents now accept a dpiMultiplier parameter. this allows the caller
   to specify an anti-alias multiplier that the PDF engine will use when it
   draws raster objects into the device context.
 - instead of creating a Skia PDF device, we now create a Skia PDF document
   and use its canvas to draw on
 - fix the calculation of DPI so it uses the diagonal length of the media
   instead of averaging the horizontal and vertical length.