Give preference to the RenderWidgetHostView's window when focusing the widget
By default, DesktopNativeWidgetAura activates the aura::Window for the root
view in the widget (if there are no stored focused views).  This commit allows
the WidgetDelegate to provide an alternate activation window.

In the case of blpwtk2::WebViewImpl, we will return the window for the current
RenderWidgetHostView.  The effect of this is that the RenderView will get
logical focus when the widget gains keyboard focus.