Propagate disablement of cache offsets across frame boundaries
When transforming a local rect to a invalidation container rect,
RenderBox::mapRectToPaintInvalidationBacking checks the paint invalidation
state to determine if it can use cached offsets. Cached offsets are enabled
by default but they are disabled if any of the following conditions are met:

1. The layout object has columns
2. The layout object has a transform property
3. The layout object has a reflection
4. The layout object is in flipped-blocks writing mode

These conditions are checked by
LayoutObject::supportsPaintInvalidationStateCachedOffsets and based on the
return value of this function, the constructor of PaintInvalidationState
disables the cached offsets mode.

The problem is that when FrameView::invalidateTreeIfNeededRecursive invalidates
the layout tree recursively, it first performs the invalidation check of all
children under a specific frame and then iterates into the child frames. This
prevents the propagation of the cached offsets mode to the child frames.

This commit checks the parent render layers of the child frame and if any layer
contains a style that should disable cached offsets, it disables it for the
child frame. It also propagates the value down to the inner child frames.