Always DisableSwapUntilResize on window size change
This achieves the same effect as [1], where the real cleanup of the
'need_synchronous_paint_' flag happened.  This commit is a temporary measure
until we pull that update.

This fixes an issue where content inside blpwtk2 WebViews "jiggle" during
resize.  Note that we don't get the Windows begin/end resizing events in
blpwtk2, so 'need_synchronous_paint_' was never being set.

[1] https://codereview.chromium.org/1148093007