Add WebView::stopFind
Previously, the WebView::find method would stop the find if the specified
string was empty.  This commit changes that by adding a stopFind method to make
it more explicit that we are stopping the find.  It also allows for a flag to
specify whether or not we want to preserve the selection of the active search
item.  Also, fix the way we handle overflow of the find request id.