Don't assume the transform will always enlarge the clip rect
Prior to this commit, the unscaled document rect was mapped to the root layer
with the transform applied and it was declared to be the clip rect. This works
only if the transform up scales the element size. If the transform down scales
the element size, the clip rect would be shorter than it needed to be. This
change ensures that the union of the two (transformed and un-transformed)
mapping is used as the clip rect.