Create an IO thread for ProcessClientImpl if necessary
Previously, when using the RENDERER_MAIN thread-mode, the IO thread that was
used by ProcessClientImpl would be the same one that was used by the in-process
renderer thread.  However, when the in-process renderer is disabled, this IO
thread does not exist.

This commit fixes this issue by creating a separate IO thread that is used
whenever the in-process renderer is disabled.

Note that the "RENDERER_MAIN" thread-mode is a bit of a misnomer when the
in-process renderer is disabled.  Something like "NOT_ORIGINAL" would be more
appropriate in this case.