Remove blpwtk2::WebView::drawContentsToDevice
The functionality provided by drawContentsToDevice can be performed by
the embedder, so there is no need for it to be implemented in the
toolkit.  Since the destination region no longer need to be specified as
a rectangular region and can be specified as a size, the DrawParams
struct is simplified so the destination size is specified instead of the
rect.

In addition to changing destination rect to destination size, the type
of the width and height are are float instead of int.  This will allow
the embedder to specify the dimension of the document in partial points.
This will prevent round-off errors in certain cases.