Simplify changes for the removeIMEUnderlineColor branch
The purpose of this branch was to make the IME underline color the same as the
text color, so the 'color' property of the CompositionUnderline object was not
needed, and was removed.

Due to changes upstream [1] in the way composition underlines are rendered,
this caused a lot of failures when updating to v46.  To simplify future
updates, the changes on this branch are being reverted (restoring the 'color'
property), but the 'paintCompositionUnderline' method has been changed to just
ignore it.

[1] https://codereview.chromium.org/1355503003