blpwtk2: Allow a desktop window to serve as a streaming media source
The Chrome browser's content API includes the capability to establish a desktop
window as a streaming media source (which can then be displayed in <video>
elements, etc.). Chrome only exposes this to JavaScript as an extension
(chrome.desktopCapture) which includes a lot of things we don't need, such as a
Chrome-specific user interface for picking a desktopwindow (besides this,
Chrome's extensions have all been removed from chromium.bb).

This commit adds a public function to blpwtk2::Toolkit that allows an
application to create a new streaming media source from a NativeView reference
(in our case, an HWND). The implementation of this is fairly similar to parts
of Chrome's: a registry (blpwtk2::DesktopStreamsRegistry) creates a
content::MediaStreamDevice from the window handle and associates it with a
string of random characters that can later be supplied to the
navigator.webkitGetUserMedia() function. Entries are removed from the registry
after they are consumed by navigator.webkitGetUserMedia() or after some set
period of time (10 seconds currently).