Avoid early-out when painting selected empty lines
This issue is being tracked upstream here:
    https://code.google.com/p/chromium/issues/detail?id=568663

Empty lines have an empty overflow rect, so the paint routine exits early
because it doesn't intersect the paint rect.

This commit expands the overflow rect using the same logic that is used in
InlineTextBox::localSelectionRect, where the selection rect is expanded by the
amount returned by newlineSpaceWidth().