Highlight table cells when the contents are fully selected
This commit takes a bit from the 'm_column' count to add a flag indicating
whether a table cell is fully selected (i.e. all children inside are selected).
This flag is updated whenever the table cell's selection state is set.

When a table cell is fully selected, we will fill it with its selection color
instead of using the background color.

The change in LayoutView.cpp is needed so that the fully-selected table cells
are invalidated when the selection is cleared.  The change in
InlineTextBoxPainter.cpp is needed so that the selection highlight is not drawn
twice (once for the containing table cell, and once again for the text).