pdfium: Avoid really shutting down SDK
Calling InitializeSDK, followed by ShutdownSDK, followed by InitializeSDK again
causes an assertion failure in FXJS_Initialize.  This is because FXJS_Release
is never called, and the isolate in the second initialization is different from
the first one.

One solution would be to call FXJS_Release in FPDF_DestroyLibrary, so that the
isolate pointer is reset.  However, it's not clear to me what other
side-effects this will have.  Also, this code has changed significantly in
newer versions of chromium (to make pdfium->v8 dependency optional).

In this commit, we instead just avoid shutting down the SDK altogether.  It
seems better to just keep it alive for the duration of the process (in the case
of utility processes, this is a very short time anyway).  The only thing
ShutdownSDK does now is exit the isolate.  When InitializeSDK is called for the
second time, we will just re-enter the isolate, and return immediately.