Add blpwtk2::WebView::setBackgroundColor()
Adds a function to set the background color of the blpwtk::WebView.
This color will be exposed if the HTML element's background color has
a non-opaque alpha channel.

The primary use for this is to set a completely transparent background so
that the webview can be used to render 'shaped' windows.  Up until now, this
had been achieved by calling blpwtk2::WebViewCreateParams::setIsTransparent(),
so that the webview would have a fully transparent background for its entire
lifetime, but we now require this more flexible mechanism.  Therefore this
commit also removes blpwtk2::WebViewCreateParams::setIsTransparent().