Introduce blpwtk2::WebContentSettingsDelegate
blink provides the abstract class 'blink::WebContentSettingsClient' that
allows an embedder to set various content permissions for a 'blink::WebFrame'
(provided that that frame is 'local').  These changes essentially start to
expose some of the capabilities of this class to applications that use blpwtk2
for embedding via the abstract class 'blpwtk2::WebContentSettingsDelegate'.

This class currently provides two of the functions of the blink class,
'blpwtk2::WebContentSettingsDelegate::allowDisplayingInsecureContent()' and
'blpwtk2::WebContentSettingsDelegate::allowRunningInsecureContent()', which
will allow a frame loaded from a secure origin to present content from an
insecure origin.

'blpwtk2::WebFrame::setContentSettingsDelegate()' is added to specify the
delegate for a frame after it has been loaded.