Provide a way to use 'CLIPFORMAT' indices to access drop event data
Adds two functions to ui::OSExchangeData::Provider: EnumerateCustomData(), which will retrieve a list of all of the FORMATETC objects available on the exchange object, and GetCustomData(), which will retrieve the data associated with a particular FORMATETC object.

Extends the function PrepareDropData(), used by WebContentsViewAura to begin transfering data from OSExchangeData object to blink, to use these two new functions to iterate over every available FORMATETC and make it available with the key "blp_N" (where N is the numeric ID of the format object).